배포

See merge request s12-webmobile2-sub1/S12P11A607!140

+++ gagunokuga/gagunokuga_back/src/main/java/com/example/gagunokuga_back/furniture/dto/FurnitureListResponse.java
@@ -10,6 +10,6 @@ import java.util.List;
 @Builder
 @ToString
 public class FurnitureListResponse {
-    private Long totalPages;
+    private int totalPages;
     private List<FurnitureResponse> furnitures;
 }
